feat(typing): improves
