+++ src/__tests__/index.test.ts
@@ -1,7 +1,8 @@
 import pkg from '../../package.json'
 
 describe('package.json', () => {
-  it(`has no dependencies`, () => {
-    expect(Object.keys(pkg.dependencies).length).toBe(0)
+  it(`has only one dependency which is for support react-native`, () => {
+    expect('react-native-svg' in pkg.dependencies).toBe(true)
+    expect(Object.keys(pkg.dependencies).length).toBe(1)
   })
 })
