+++ src/native/package.json
@@ -1,7 +1,8 @@
 {
   "name": "react-content-loader/native",
   "private": true,
-  "main": "./native/react-content-loader.native.cjs.js",
-  "module": "./native/react-content-loader.native.es.js",
-  "jsnext:main": "./native/react-content-loader.native.es.js"
+  "main": "./react-content-loader.native.cjs.js",
+  "module": "./react-content-loader.native.es.js",
+  "jsnext:main": "./react-content-loader.native.es.js",
+  "types": "./native/index.d.ts"
 }
