+++ tsconfig.json
@@ -1,15 +1,16 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "module": "commonjs",
     "jsx": "react",
+    "lib": ["es6"],
     "declaration": true,
     "declarationDir": "./dist/types",
     "declarationMap": true,
     "noEmit": true,
     "sourceMap": true,
     "esModuleInterop": true,
-    "noImplicitAny": false
+    "noImplicitAny": false,
+    "resolveJsonModule": true
   },
   "include": ["./src/**/*"],
   "exclude": ["node_modules", "src/__tests__"]
